Edit20240624

+++ AuthAppDemoService/Basics/Dtos/DtoClasses.cs
@@ -102,7 +102,7 @@ namespace AuthAppDemoService.Basics.Dtos
         /// <summary>
         /// Total count of Items.
         /// </summary>
-        public int TotalCount { get; set; }
+        public long TotalCount { get; set; }
 
         /// <summary>
         /// Creates a new <see cref="PagedResultDto{T}"/> object.
@@ -117,7 +117,7 @@ namespace AuthAppDemoService.Basics.Dtos
         /// </summary>
         /// <param name="totalCount">Total count of Items</param>
         /// <param name="items">List of items in current page</param>
-        public PagedResultDto(int totalCount, IReadOnlyList<T> items)
+        public PagedResultDto(long totalCount, IReadOnlyList<T> items)
             : base(items)
         {
             TotalCount = totalCount;
